+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+* The blanket implementations of `Fmt` for references now apply to `!Sized` types too.
+
 ## 0.1.0 (2023-10-07)
 
 Initial public release.